requirements.txt
paho-mqtt
pymongo
paho-mqtt pymongo
fastapi uvicorn
